ビルドエラー解消_dbz

+++ DB/DBZ/standard/business/src/main/java/jp/co/ndensan/reams/db/dbz/business/core/ShikakuShutokuJogaisha.java
@@ -15,7 +15,6 @@ import jp.co.ndensan.reams.uz.uza.biz.LasdecCode;
 import jp.co.ndensan.reams.uz.uza.biz.ShikibetsuCode;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleDate;
 import jp.co.ndensan.reams.uz.uza.lang.RString;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import jp.co.ndensan.reams.uz.uza.util.db.EntityDataState;
 
 /**
@@ -34,7 +33,7 @@ public class ShikakuShutokuJogaisha extends ModelBase<ShikakuShutokuJogaishaIden
      * @param 履歴番号 履歴番号
      */
     public ShikakuShutokuJogaisha(ShikibetsuCode 識別コード,
-            Decimal 履歴番号) {
+            int 履歴番号) {
         requireNonNull(識別コード, UrSystemErrorMessages.値がnull.getReplacedMessage("識別コード"));
         requireNonNull(履歴番号, UrSystemErrorMessages.値がnull.getReplacedMessage("履歴番号"));
         this.entity = new DbT1009ShikakuShutokuJogaishaEntity();
@@ -88,7 +87,7 @@ public class ShikakuShutokuJogaisha extends ModelBase<ShikakuShutokuJogaishaIden
      *
      * @return 履歴番号
      */
-    public Decimal get履歴番号() {
+    public int get履歴番号() {
         return entity.getRirekiNo();
     }
