+++ DB/DBZ/standard/business/src/main/java/jp/co/ndensan/reams/db/dbz/business/core/ShikakuShutokuJogaishaBuilder.java
@@ -12,7 +12,6 @@ import jp.co.ndensan.reams.uz.uza.biz.LasdecCode;
 import jp.co.ndensan.reams.uz.uza.biz.ShikibetsuCode;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleDate;
 import jp.co.ndensan.reams.uz.uza.lang.RString;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
 
 /**
  * {@link ShikakuShutokuJogaisha}の編集を行うビルダークラスです。
@@ -58,7 +57,7 @@ public class ShikakuShutokuJogaishaBuilder {
      * @param 履歴番号 履歴番号
      * @return {@link ShikakuShutokuJogaishaBuilder}
      */
-    public ShikakuShutokuJogaishaBuilder set履歴番号(Decimal 履歴番号) {
+    public ShikakuShutokuJogaishaBuilder set履歴番号(int 履歴番号) {
         requireNonNull(履歴番号, UrSystemErrorMessages.値がnull.getReplacedMessage("履歴番号"));
         entity.setRirekiNo(履歴番号);
         return this;
